--- src/main/java/com/example/tubesRPL/User/LoginController.java
@@ -1,40 +1,0 @@
-package com.example.tubesRPL.User;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-import jakarta.servlet.http.HttpSession;
-
-@Controller
-public class LoginController {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private BCryptPasswordEncoder passwordEncoder;
-
-    @GetMapping("/login")
-    public String showLoginForm() {
-        return "/login";
-    }
-
-    @GetMapping("/register")
-    public String showRegisterForm() {
-        return "/register";
-    }
-
-    @PostMapping
-    public String login(@RequestParam String email, @RequestParam String password, HttpSession session) {
-        User user = userRepository.findByEmail(email);
-        if (user != null && passwordEncoder.matches(password, user.getPassword())) {
-            session.setAttribute("loggedInUser", user);
-            return "Welcome " + user.getName() + "!";
-        }
-        return "Invalid credentials!";
-    }
-}
