--- src/main/java/com/example/tubesRPL/User/UserRepository.java
@@ -1,42 +1,0 @@
-package com.example.tubesRPL.User;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Repository;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.List;
-
-@Repository
-public class UserRepository {
-
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-
-    @Autowired
-    private BCryptPasswordEncoder passwordEncoder;
-
-    public void save(User user) {
-        String encryptedPassword = passwordEncoder.encode(user.getPassword());
-        String sql = "INSERT INTO users (name, email, password, role) VALUES (?, ?, ?, ?)";
-        jdbcTemplate.update(sql, user.getName(), user.getEmail(), encryptedPassword, user.getRole());
-    }
-
-    public User findByEmail(String email) {
-        String sql = "SELECT * FROM users WHERE email = ?";
-        List<User> users = jdbcTemplate.query(sql, this::mapRowToUser, email);
-        return users.isEmpty() ? null : users.get(0);
-    }
-
-    private User mapRowToUser(ResultSet rs, int rowNum) throws SQLException {
-        User user = new User();
-        user.setId(rs.getInt("id"));
-        user.setName(rs.getString("name"));
-        user.setEmail(rs.getString("email"));
-        user.setPassword(rs.getString("password"));
-        user.setRole(rs.getString("role"));
-        return user;
-    }
-}
